+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `@Platform(pattern=...)` annotation value to allow matching with regular expressions as well
  * Allow `Parser` to consider function pointers declared with `using` but without indirections
  * Let `Parser` map to annotations whole expressions starting with the `__attribute__` keyword
  * Fix `Parser` sometimes failing to create template instances with default arguments ([issue #478](https://github.com/bytedeco/javacpp/issues/478))